Android: Merge NativeActivity into PpssppActivity
